Allow deleting a single category
