+++ src/applications/content-categories-app/categories/categories-list/categories-list.component.html
