+++ src/applications/content-categories-app/categories/categories-table/categories-table.component.scss
