+++ src/applications/content-categories-app/categories/pipes/prime-table-sort-direction.pipe.ts
@@ -1,6 +1,6 @@
-import { Pipe, PipeTransform } from '@angular/core';
-import { AppLocalization } from '@kaltura-ng/kaltura-common';
-import { SortDirection } from "../../categories/categories.service";
+import {Pipe, PipeTransform} from '@angular/core';
+import {AppLocalization} from '@kaltura-ng/kaltura-common';
+import {SortDirection} from '../../categories/categories.service';
 
 
 @Pipe({name: 'kPrimeTableSortDirection'})
@@ -8,14 +8,12 @@ export class PrimeTableSortDirectionPipe implements PipeTransform {
 	constructor(private appLocalization: AppLocalization) {
 	}
 
-	transform(value: SortDirection,): number {
-
-    switch (value)
-	{
-		case SortDirection.Asc:
-			return 1;
-		case SortDirection.Desc:
-			return -1;
-	}
+	transform(value: SortDirection): number {
+    switch (value) {
+      case SortDirection.Asc:
+        return 1;
+      case SortDirection.Desc:
+        return -1;
+    }
   }
 }
