+++ src/applications/content-categories-app/content-categories-app.routes.ts
@@ -1,13 +1,13 @@
-import { CategoryCanDeactivate } from 'applications/content-categories-app/category/category-can-deactivate.service';
-import { CategoryEntitlementsComponent } from './category/category-entitlements/category-entitlements.component';
-import { CategoryMetadataComponent } from './category/category-metadata/category-metadata.component';
-import { CategoryWidgetKeys } from './category/category-widget-keys';
-import { CategorySubcategoriesComponent } from './category/category-subcategories/category-subcategories.component';
-import { Route } from '@angular/router';
+import {CategoryCanDeactivate} from 'applications/content-categories-app/category/category-can-deactivate.service';
+import {CategoryEntitlementsComponent} from './category/category-entitlements/category-entitlements.component';
+import {CategoryMetadataComponent} from './category/category-metadata/category-metadata.component';
+import {CategoryWidgetKeys} from './category/category-widget-keys';
+import {CategorySubcategoriesComponent} from './category/category-subcategories/category-subcategories.component';
+import {Route} from '@angular/router';
 
-import { ContentCategoriesComponent } from './content-categories.component';
-import { CategoriesListComponent } from './categories/categories-list.component';
-import { CategoryComponent } from './category/category.component';
+import {ContentCategoriesComponent} from './content-categories.component';
+import {CategoriesListComponent} from './categories/categories-list/categories-list.component';
+import {CategoryComponent} from './category/category.component';
 
 export const routing: Route[] = [
 	{
