+++ src/i18n/en.json
@@ -246,15 +246,19 @@
         },
         "deleteCategories": "Delete Categories",
         "editWarning": "Some of the selected categories were created by another application and deleting them will effect this application or site.\n",
-        "confirmDeleteWithSubCategories": "The category will be deleted with its sub-categories. Do you want to continue?\n {{0}}",
-        "confirmDeleteSingle": "Are you sure you want to delete this category?\n {{0}}",
+        "deleteWithEditWarningTags": "This category was created by another application and deleting it will effect this application or site.\n Do you want to continue?",
+        "confirmDeleteWithSubCategories": "The category will be deleted with its sub-categories. Do you want to continue?",
+        "confirmDeleteSingle": "Are you sure you want to delete this category?",
         "confirmDeleteMultipleWithSubCategories": "Categories will be deleted with their sub-categories. Do you want to continue?\n {{0}}",
         "confirmDeleteMultiple": "Are you sure you want to delete these categories?\n {{0}}",
         "deleted": "Category deleted successfully",
         "categoryId": "Category ID: {{0}}",
         "noRestriction": "No Restriction",
         "requiresAuthentication": "Requires Authentication",
-        "private": "Private"
+        "private": "Private",
+        "errors" : {
+          "categoryCouldNotBeDeleted" : "Category could not be deleted"
+        }
       },
       "addNewCategory": {
         "newCategory": "New Category",
